refactor: make DatabaseUtil application에서 구현한 method를 DataBaseUtil로 이동
