+++ src/main/java/gift/utility/DataBaseUtil.java
@@ -1,0 +1,33 @@
+package gift.utility;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Component;
+
+@Component
+public class DataBaseUtil {
+
+    private JdbcTemplate jdbcTemplate;
+
+    @Autowired
+    public DataBaseUtil(JdbcTemplate jdbcTemplate){
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public void dropProductTable(){
+        var sql="DROP TABLE products IF EXISTS";
+        jdbcTemplate.execute(sql);
+    }
+
+    public void createProductTable(){
+        var sql= """
+            create table products(
+                id bigint,
+                name varchar(255),
+                price int,
+                image_url varchar(255)
+            )
+            """;
+        jdbcTemplate.execute(sql);
+    }
+}
